refactor: extract campaign persistence store
