+++ backend/src/users.rs
@@ -1,10 +1,7 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use sqlx::{sqlite::SqlitePoolOptions, Row, SqlitePool};
-use std::{
-    collections::{BTreeMap, BTreeSet},
-    path::Path,
-};
+use std::{collections::BTreeMap, path::Path};
 use thiserror::Error;
 
 pub(crate) const MAX_DISPLAY_NAME_LENGTH: usize = 48;
@@ -74,36 +71,6 @@ impl UserStore {
                 user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
                 expires_at INTEGER NOT NULL,
                 created_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-            )
-            "#,
-        )
-        .execute(&mut *tx)
-        .await?;
-
-        sqlx::query(
-            r#"
-            CREATE TABLE IF NOT EXISTS campaign_presets (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                display_name TEXT NOT NULL,
-                room_slug TEXT NOT NULL UNIQUE,
-                default_split_room_names TEXT NOT NULL DEFAULT '[]',
-                is_archived INTEGER NOT NULL DEFAULT 0,
-                created_by_user_id INTEGER REFERENCES users(id) ON DELETE SET NULL,
-                created_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-                updated_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-            )
-            "#,
-        )
-        .execute(&mut *tx)
-        .await?;
-
-        sqlx::query(
-            r#"
-            CREATE TABLE IF NOT EXISTS campaign_members (
-                campaign_id INTEGER NOT NULL REFERENCES campaign_presets(id) ON DELETE CASCADE,
-                user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-                game_role TEXT NOT NULL CHECK (game_role IN ('gamemaster', 'player')),
-                PRIMARY KEY (campaign_id, user_id)
             )
             "#,
         )
@@ -508,227 +475,6 @@ impl UserStore {
         Ok(())
     }
 
-    pub async fn list_campaigns_for_user(
-        &self,
-        user_id: i64,
-        is_admin: bool,
-    ) -> Result<Vec<CampaignPreset>, StoreError> {
-        let rows = if is_admin {
-            sqlx::query(
-                r#"
-                SELECT id, display_name, room_slug, default_split_room_names,
-                       is_archived, created_at, updated_at
-                FROM campaign_presets
-                WHERE is_archived = 0
-                ORDER BY display_name COLLATE NOCASE ASC
-                "#,
-            )
-            .fetch_all(&self.pool)
-            .await?
-        } else {
-            sqlx::query(
-                r#"
-                SELECT campaign_presets.id, campaign_presets.display_name,
-                       campaign_presets.room_slug,
-                       campaign_presets.default_split_room_names,
-                       campaign_presets.is_archived,
-                       campaign_presets.created_at,
-                       campaign_presets.updated_at
-                FROM campaign_presets
-                INNER JOIN campaign_members
-                    ON campaign_members.campaign_id = campaign_presets.id
-                WHERE campaign_members.user_id = ?
-                  AND campaign_presets.is_archived = 0
-                ORDER BY campaign_presets.display_name COLLATE NOCASE ASC
-                "#,
-            )
-            .bind(user_id)
-            .fetch_all(&self.pool)
-            .await?
-        };
-
-        let mut campaigns = Vec::with_capacity(rows.len());
-        for row in rows {
-            campaigns.push(self.campaign_from_row(row).await?);
-        }
-        Ok(campaigns)
-    }
-
-    pub async fn list_managed_campaigns(
-        &self,
-        user_id: i64,
-        is_admin: bool,
-    ) -> Result<Vec<CampaignPreset>, StoreError> {
-        let rows = if is_admin {
-            sqlx::query(
-                r#"
-                SELECT id, display_name, room_slug, default_split_room_names,
-                       is_archived, created_at, updated_at
-                FROM campaign_presets
-                ORDER BY is_archived ASC, display_name COLLATE NOCASE ASC
-                "#,
-            )
-            .fetch_all(&self.pool)
-            .await?
-        } else {
-            sqlx::query(
-                r#"
-                SELECT campaign_presets.id, campaign_presets.display_name,
-                       campaign_presets.room_slug,
-                       campaign_presets.default_split_room_names,
-                       campaign_presets.is_archived,
-                       campaign_presets.created_at,
-                       campaign_presets.updated_at
-                FROM campaign_presets
-                INNER JOIN campaign_members
-                    ON campaign_members.campaign_id = campaign_presets.id
-                WHERE campaign_members.user_id = ?
-                  AND campaign_members.game_role = 'gamemaster'
-                ORDER BY campaign_presets.is_archived ASC,
-                         campaign_presets.display_name COLLATE NOCASE ASC
-                "#,
-            )
-            .bind(user_id)
-            .fetch_all(&self.pool)
-            .await?
-        };
-
-        let mut campaigns = Vec::with_capacity(rows.len());
-        for row in rows {
-            campaigns.push(self.campaign_from_row(row).await?);
-        }
-        Ok(campaigns)
-    }
-
-    pub async fn create_campaign(
-        &self,
-        creator_user_id: i64,
-        input: CampaignInput,
-    ) -> Result<CampaignPreset, StoreError> {
-        let room_slug = normalize_room_slug(&input.room_slug);
-        let split_rooms =
-            serde_json::to_string(&sanitize_split_room_names(&input.default_split_room_names))?;
-        let mut tx = self.pool.begin().await?;
-        let result = sqlx::query(
-            r#"
-            INSERT INTO campaign_presets (
-                display_name, room_slug, default_split_room_names, is_archived, created_by_user_id
-            ) VALUES (?, ?, ?, ?, ?)
-            "#,
-        )
-        .bind(input.display_name.trim())
-        .bind(&room_slug)
-        .bind(split_rooms)
-        .bind(if input.is_archived { 1_i64 } else { 0_i64 })
-        .bind(creator_user_id)
-        .execute(&mut *tx)
-        .await;
-
-        let campaign_id = match result {
-            Ok(record) => record.last_insert_rowid(),
-            Err(sqlx::Error::Database(db_err)) if db_err.is_unique_violation() => {
-                return Err(StoreError::CampaignSlugAlreadyExists(room_slug));
-            }
-            Err(err) => return Err(StoreError::Sqlx(err)),
-        };
-        replace_campaign_members(&mut tx, campaign_id, &input).await?;
-        tx.commit().await?;
-        self.find_campaign_by_id(campaign_id)
-            .await?
-            .ok_or(StoreError::CampaignNotFound(campaign_id))
-    }
-
-    pub async fn update_campaign(
-        &self,
-        campaign_id: i64,
-        input: CampaignInput,
-    ) -> Result<CampaignPreset, StoreError> {
-        let room_slug = normalize_room_slug(&input.room_slug);
-        let split_rooms =
-            serde_json::to_string(&sanitize_split_room_names(&input.default_split_room_names))?;
-        let mut tx = self.pool.begin().await?;
-        let result = sqlx::query(
-            r#"
-            UPDATE campaign_presets
-            SET display_name = ?, room_slug = ?, default_split_room_names = ?,
-                is_archived = ?, updated_at = CURRENT_TIMESTAMP
-            WHERE id = ?
-            "#,
-        )
-        .bind(input.display_name.trim())
-        .bind(&room_slug)
-        .bind(split_rooms)
-        .bind(if input.is_archived { 1_i64 } else { 0_i64 })
-        .bind(campaign_id)
-        .execute(&mut *tx)
-        .await;
-
-        match result {
-            Ok(result) if result.rows_affected() == 0 => {
-                return Err(StoreError::CampaignNotFound(campaign_id));
-            }
-            Ok(_) => {}
-            Err(sqlx::Error::Database(db_err)) if db_err.is_unique_violation() => {
-                return Err(StoreError::CampaignSlugAlreadyExists(room_slug));
-            }
-            Err(err) => return Err(StoreError::Sqlx(err)),
-        }
-        replace_campaign_members(&mut tx, campaign_id, &input).await?;
-        tx.commit().await?;
-        self.find_campaign_by_id(campaign_id)
-            .await?
-            .ok_or(StoreError::CampaignNotFound(campaign_id))
-    }
-
-    pub async fn archive_campaign(&self, campaign_id: i64) -> Result<(), StoreError> {
-        let result = sqlx::query(
-            "UPDATE campaign_presets SET is_archived = 1, updated_at = CURRENT_TIMESTAMP WHERE id = ?",
-        )
-        .bind(campaign_id)
-        .execute(&self.pool)
-        .await?;
-        if result.rows_affected() == 0 {
-            return Err(StoreError::CampaignNotFound(campaign_id));
-        }
-        Ok(())
-    }
-
-    pub async fn find_campaign_by_room_slug(
-        &self,
-        room_slug: &str,
-    ) -> Result<Option<CampaignPreset>, StoreError> {
-        let row = sqlx::query(
-            r#"
-            SELECT id, display_name, room_slug, default_split_room_names,
-                   is_archived, created_at, updated_at
-            FROM campaign_presets
-            WHERE room_slug = ?
-            "#,
-        )
-        .bind(normalize_room_slug(room_slug))
-        .fetch_optional(&self.pool)
-        .await?;
-        match row {
-            Some(row) => Ok(Some(self.campaign_from_row(row).await?)),
-            None => Ok(None),
-        }
-    }
-
-    pub async fn campaign_role_for_user(
-        &self,
-        campaign_id: i64,
-        user_id: i64,
-    ) -> Result<Option<GameRole>, StoreError> {
-        let role = sqlx::query_scalar::<_, String>(
-            "SELECT game_role FROM campaign_members WHERE campaign_id = ? AND user_id = ?",
-        )
-        .bind(campaign_id)
-        .bind(user_id)
-        .fetch_optional(&self.pool)
-        .await?;
-        role.map(GameRole::from_db_value).transpose()
-    }
-
     pub async fn users_are_active(&self, user_ids: &[i64]) -> Result<bool, StoreError> {
         for user_id in user_ids {
             let Some(user) = self.find_user_by_id(*user_id).await? else {
@@ -739,82 +485,6 @@ impl UserStore {
             }
         }
         Ok(true)
-    }
-
-    pub async fn user_can_manage_campaign(
-        &self,
-        campaign_id: i64,
-        user_id: i64,
-        is_admin: bool,
-    ) -> Result<bool, StoreError> {
-        if is_admin {
-            return Ok(true);
-        }
-        let count = sqlx::query_scalar::<_, i64>(
-            r#"
-            SELECT COUNT(*) FROM campaign_members
-            WHERE campaign_id = ? AND user_id = ? AND game_role = 'gamemaster'
-            "#,
-        )
-        .bind(campaign_id)
-        .bind(user_id)
-        .fetch_one(&self.pool)
-        .await?;
-        Ok(count > 0)
-    }
-
-    pub(crate) async fn find_campaign_by_id(
-        &self,
-        campaign_id: i64,
-    ) -> Result<Option<CampaignPreset>, StoreError> {
-        let row = sqlx::query(
-            r#"
-            SELECT id, display_name, room_slug, default_split_room_names,
-                   is_archived, created_at, updated_at
-            FROM campaign_presets WHERE id = ?
-            "#,
-        )
-        .bind(campaign_id)
-        .fetch_optional(&self.pool)
-        .await?;
-        match row {
-            Some(row) => Ok(Some(self.campaign_from_row(row).await?)),
-            None => Ok(None),
-        }
-    }
-
-    async fn campaign_from_row(
-        &self,
-        row: sqlx::sqlite::SqliteRow,
-    ) -> Result<CampaignPreset, StoreError> {
-        let campaign_id = row.get::<i64, _>("id");
-        let member_rows = sqlx::query(
-            "SELECT user_id, game_role FROM campaign_members WHERE campaign_id = ? ORDER BY user_id",
-        )
-        .bind(campaign_id)
-        .fetch_all(&self.pool)
-        .await?;
-        let mut gamemaster_user_ids = Vec::new();
-        let mut player_user_ids = Vec::new();
-        for member in member_rows {
-            match member.get::<String, _>("game_role").as_str() {
-                "gamemaster" => gamemaster_user_ids.push(member.get("user_id")),
-                "player" => player_user_ids.push(member.get("user_id")),
-                other => return Err(StoreError::InvalidRoleValue(other.to_string())),
-            }
-        }
-        let split_room_json = row.get::<String, _>("default_split_room_names");
-        Ok(CampaignPreset {
-            id: campaign_id,
-            display_name: row.get("display_name"),
-            room_slug: row.get("room_slug"),
-            gamemaster_user_ids,
-            player_user_ids,
-            default_split_room_names: serde_json::from_str(&split_room_json)?,
-            is_archived: row.get::<i64, _>("is_archived") != 0,
-            created_at: row.get("created_at"),
-            updated_at: row.get("updated_at"),
-        })
     }
 
     pub async fn find_user_by_id(&self, user_id: i64) -> Result<Option<AuthUser>, StoreError> {
@@ -914,29 +584,6 @@ pub struct UserPatch {
     pub is_active: Option<bool>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct CampaignPreset {
-    pub id: i64,
-    pub display_name: String,
-    pub room_slug: String,
-    pub gamemaster_user_ids: Vec<i64>,
-    pub player_user_ids: Vec<i64>,
-    pub default_split_room_names: Vec<String>,
-    pub is_archived: bool,
-    pub created_at: String,
-    pub updated_at: String,
-}
-
-#[derive(Debug, Clone)]
-pub struct CampaignInput {
-    pub display_name: String,
-    pub room_slug: String,
-    pub gamemaster_user_ids: Vec<i64>,
-    pub player_user_ids: Vec<i64>,
-    pub default_split_room_names: Vec<String>,
-    pub is_archived: bool,
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 pub enum PlatformRole {
@@ -999,10 +646,7 @@ pub enum StoreError {
     InvalidEmail(String),
     #[error("email already exists: {0}")]
     EmailAlreadyExists(String),
-    #[error("campaign not found: {0}")]
-    CampaignNotFound(i64),
-    #[error("campaign room slug already exists: {0}")]
-    CampaignSlugAlreadyExists(String),
+
     #[error("google subject mismatch for {0}; existing subject is {1}")]
     GoogleSubjectMismatch(String, String),
     #[error("cannot remove, demote, or disable the last active admin")]
@@ -1011,60 +655,9 @@ pub enum StoreError {
     InvalidRoleValue(String),
     #[error("database error: {0}")]
     Sqlx(#[from] sqlx::Error),
-    #[error("invalid stored JSON: {0}")]
-    Json(#[from] serde_json::Error),
+
     #[error("filesystem error: {0}")]
     Io(#[from] std::io::Error),
-}
-
-async fn replace_campaign_members(
-    tx: &mut sqlx::Transaction<'_, sqlx::Sqlite>,
-    campaign_id: i64,
-    input: &CampaignInput,
-) -> Result<(), StoreError> {
-    sqlx::query("DELETE FROM campaign_members WHERE campaign_id = ?")
-        .bind(campaign_id)
-        .execute(&mut **tx)
-        .await?;
-    let mut inserted_user_ids = BTreeSet::new();
-    for user_id in &input.gamemaster_user_ids {
-        if !inserted_user_ids.insert(*user_id) {
-            continue;
-        }
-        sqlx::query(
-            "INSERT INTO campaign_members (campaign_id, user_id, game_role) VALUES (?, ?, 'gamemaster')",
-        )
-        .bind(campaign_id)
-        .bind(user_id)
-        .execute(&mut **tx)
-        .await?;
-    }
-    for user_id in &input.player_user_ids {
-        if !inserted_user_ids.insert(*user_id) {
-            continue;
-        }
-        sqlx::query(
-            "INSERT INTO campaign_members (campaign_id, user_id, game_role) VALUES (?, ?, 'player')",
-        )
-        .bind(campaign_id)
-        .bind(user_id)
-        .execute(&mut **tx)
-        .await?;
-    }
-    Ok(())
-}
-
-fn normalize_room_slug(value: &str) -> String {
-    value.trim().to_lowercase()
-}
-
-fn sanitize_split_room_names(values: &[String]) -> Vec<String> {
-    values
-        .iter()
-        .map(|value| value.trim())
-        .filter(|value| !value.is_empty())
-        .map(ToString::to_string)
-        .collect()
 }
 
 pub fn build_bootstrap_users(
@@ -1302,155 +895,6 @@ mod tests {
 
         store.delete_session("session-1").await.unwrap();
         assert!(store.get_session_user("session-1").await.unwrap().is_none());
-    }
-
-    #[tokio::test]
-    async fn campaign_presets_persist_membership_and_archive_state() {
-        let store = UserStore::connect("sqlite::memory:").await.unwrap();
-        let users = build_bootstrap_users(
-            &["admin@example.com".to_string()],
-            &["gm@example.com".to_string()],
-            &["player@example.com".to_string()],
-        )
-        .unwrap();
-        store.seed_bootstrap_users(&users).await.unwrap();
-        let admin = store
-            .find_user_by_email("admin@example.com")
-            .await
-            .unwrap()
-            .unwrap();
-        let gm = store
-            .find_user_by_email("gm@example.com")
-            .await
-            .unwrap()
-            .unwrap();
-        let player = store
-            .find_user_by_email("player@example.com")
-            .await
-            .unwrap()
-            .unwrap();
-
-        let campaign = store
-            .create_campaign(
-                admin.id,
-                CampaignInput {
-                    display_name: "Thursday Night".to_string(),
-                    room_slug: "Thursday-Night".to_string(),
-                    gamemaster_user_ids: vec![gm.id],
-                    player_user_ids: vec![player.id],
-                    default_split_room_names: vec![" Library ".to_string()],
-                    is_archived: false,
-                },
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(campaign.room_slug, "thursday-night");
-        assert_eq!(campaign.default_split_room_names, vec!["Library"]);
-        assert_eq!(
-            store
-                .campaign_role_for_user(campaign.id, player.id)
-                .await
-                .unwrap(),
-            Some(GameRole::Player)
-        );
-        assert!(store
-            .user_can_manage_campaign(campaign.id, gm.id, false)
-            .await
-            .unwrap());
-        assert!(!store
-            .user_can_manage_campaign(campaign.id, player.id, false)
-            .await
-            .unwrap());
-        assert_eq!(
-            store
-                .list_campaigns_for_user(player.id, false)
-                .await
-                .unwrap()
-                .len(),
-            1
-        );
-
-        store.archive_campaign(campaign.id).await.unwrap();
-        assert!(store
-            .list_campaigns_for_user(player.id, false)
-            .await
-            .unwrap()
-            .is_empty());
-        assert!(
-            store
-                .find_campaign_by_room_slug("thursday-night")
-                .await
-                .unwrap()
-                .unwrap()
-                .is_archived
-        );
-    }
-
-    #[tokio::test]
-    async fn campaign_room_slugs_are_unique() {
-        let store = UserStore::connect("sqlite::memory:").await.unwrap();
-        let users = build_bootstrap_users(&[], &["gm@example.com".to_string()], &[]).unwrap();
-        store.seed_bootstrap_users(&users).await.unwrap();
-        let gm = store
-            .find_user_by_email("gm@example.com")
-            .await
-            .unwrap()
-            .unwrap();
-        let input = CampaignInput {
-            display_name: "First".to_string(),
-            room_slug: "same-room".to_string(),
-            gamemaster_user_ids: vec![gm.id],
-            player_user_ids: vec![],
-            default_split_room_names: vec![],
-            is_archived: false,
-        };
-        store.create_campaign(gm.id, input.clone()).await.unwrap();
-        let error = store.create_campaign(gm.id, input).await.unwrap_err();
-        assert!(matches!(error, StoreError::CampaignSlugAlreadyExists(_)));
-    }
-
-    #[tokio::test]
-    async fn campaign_presets_survive_database_reconnect() {
-        let path = std::env::temp_dir().join(format!(
-            "virtual-table-campaign-{}-{}.sqlite",
-            std::process::id(),
-            Utc::now().timestamp_micros()
-        ));
-        std::fs::File::create(&path).unwrap();
-        let database_url = format!("sqlite://{}", path.display());
-        let store = UserStore::connect(&database_url).await.unwrap();
-        let users = build_bootstrap_users(&[], &["gm@example.com".to_string()], &[]).unwrap();
-        store.seed_bootstrap_users(&users).await.unwrap();
-        let gm = store
-            .find_user_by_email("gm@example.com")
-            .await
-            .unwrap()
-            .unwrap();
-        store
-            .create_campaign(
-                gm.id,
-                CampaignInput {
-                    display_name: "Persistent Table".to_string(),
-                    room_slug: "persistent-table".to_string(),
-                    gamemaster_user_ids: vec![gm.id],
-                    player_user_ids: vec![],
-                    default_split_room_names: vec![],
-                    is_archived: false,
-                },
-            )
-            .await
-            .unwrap();
-        store.pool.close().await;
-
-        let reopened = UserStore::connect(&database_url).await.unwrap();
-        assert!(reopened
-            .find_campaign_by_room_slug("persistent-table")
-            .await
-            .unwrap()
-            .is_some());
-        reopened.pool.close().await;
-        std::fs::remove_file(path).unwrap();
     }
 
     #[test]
